Refactor landing page
