+++ password-manager/src/App.js
@@ -1,15 +1,15 @@
 // src/App.js
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Auth from './components/Auth';
-import Home from './components/Home';
+import LandingPage from './pages/LandingPage';
+import Home from './pages/Home';
 import PrivateRoute from './components/PrivateRoute';
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Auth />} />
+        <Route path="/" element={<LandingPage />} />
         <Route
           path="/home"
           element={
